# testdata/openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorResponse": {
        "description": "Error response",
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "TaskRequest": {
        "description": "Request body for creating or updating a task",
        "example": {
          "cron": "0 0 * * *",
          "method": "POST",
          "name": "Daily cleanup",
          "url": "https://example.com/api/cleanup"
        },
        "properties": {
          "body": {
            "nullable": true,
            "type": "string"
          },
          "callback_url": {
            "description": "URL to receive POST with execution results",
            "format": "uri",
            "nullable": true,
            "type": "string"
          },
          "cron": {
            "description": "Cron expression for recurring tasks",
            "nullable": true,
            "type": "string"
          },
          "enabled": {
            "default": true,
            "type": "boolean"
          },
          "expected_body_pattern": {
            "description": "Response body must contain this text to count as success. Empty means any body.",
            "nullable": true,
            "type": "string"
          },
          "expected_status_codes": {
            "description": "Comma-separated HTTP status codes that count as success (e.g. \"200,201\"). Empty means any 2xx.",
            "nullable": true,
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "method": {
            "default": "GET",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ],
            "type": "string"
          },
          "name": {
            "description": "Task name",
            "type": "string"
          },
          "notify_on_failure": {
            "description": "Override org-level failure notification setting. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "notify_on_recovery": {
            "description": "Override org-level recovery notification setting. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "retry_attempts": {
            "default": 3,
            "type": "integer"
          },
          "schedule_type": {
            "enum": [
              "cron",
              "once"
            ],
            "type": "string"
          },
          "run_at": {
            "description": "ISO 8601 timestamp for one-time scheduled execution",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "timeout_ms": {
            "default": 30000,
            "type": "integer"
          },
          "url": {
            "description": "Webhook URL to call",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "schedule_type"
        ],
        "title": "TaskRequest",
        "type": "object"
      },
      "TaskResponse": {
        "description": "Response containing a task",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Task"
          }
        },
        "title": "TaskResponse",
        "type": "object"
      },
      "TasksResponse": {
        "description": "Response containing a paginated list of tasks",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array"
          },
          "has_more": {
            "description": "Whether more results exist beyond this page",
            "type": "boolean"
          },
          "limit": {
            "description": "Maximum results per page",
            "type": "integer"
          },
          "offset": {
            "description": "Number of results skipped",
            "type": "integer"
          }
        },
        "title": "TasksResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "API key in format: pk_live_xxx.sk_live_yyy",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for managing scheduled tasks and webhooks.\n\n## Authentication\n\nAll API requests require a Bearer token in the Authorization header:\n\n```\nAuthorization: Bearer pk_live_xxx.sk_live_yyy\n```\n\nAPI keys can be created in the Runlater dashboard under Organization Settings.\n",
    "title": "Runlater API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "callbacks": {},
        "description": "Returns tasks for the authenticated organization with pagination. Optionally filter by lane name.",
        "operationId": "PrikkeWeb.Api.TaskController.index",
        "parameters": [
          {
            "description": "Filter by lane name. Use \"none\" to get tasks without a lane.",
            "in": "query",
            "name": "lane",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum results (1-100, default 50)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Number of results to skip (default 0)",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasksResponse"
                }
              }
            },
            "description": "Tasks list"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all tasks",
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "callbacks": {},
        "description": "Creates a new task. The task type is inferred from parameters:\n\n- Pass `cron` for a recurring task (e.g. `\"cron\": \"*/5 * * * *\"`)\n- Pass `delay` for delayed execution (e.g. `\"delay\": \"5m\"`)\n- Pass `run_at` for scheduled execution (e.g. `\"run_at\": \"2024-01-01T00:00:00Z\"`)\n- No timing parameter \u2192 immediate execution\n\nSupports idempotency: pass an `Idempotency-Key` header to prevent duplicate\ntask creation.\n",
        "operationId": "PrikkeWeb.Api.TaskController.create",
        "parameters": [
          {
            "description": "Unique key to prevent duplicate requests (valid for 24 hours)",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "description": "Task parameters",
          "required": false
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Task created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a task",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}": {
      "delete": {
        "callbacks": {},
        "description": "Deletes a task and all its execution history",
        "operationId": "PrikkeWeb.Api.TaskController.delete",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a task",
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "callbacks": {},
        "description": "Returns a single task by ID",
        "operationId": "PrikkeWeb.Api.TaskController.show",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Task"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a task",
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "callbacks": {},
        "description": "Updates an existing task",
        "operationId": "PrikkeWeb.Api.TaskController.update (2)",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "description": "Task parameters",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Task updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a task",
        "tags": [
          "Tasks"
        ]
      },
      "put": {
        "callbacks": {},
        "description": "Updates an existing task",
        "operationId": "PrikkeWeb.Api.TaskController.update",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "description": "Task parameters",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Task updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a task",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/executions": {
      "get": {
        "callbacks": {},
        "description": "Returns the execution history for a task",
        "operationId": "PrikkeWeb.Api.TaskController.executions",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum results (1-100, default 50)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionsResponse"
                }
              }
            },
            "description": "Executions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get execution history",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/trigger": {
      "post": {
        "callbacks": {},
        "description": "Triggers a task to run immediately, creating a pending execution",
        "operationId": "PrikkeWeb.Api.TaskController.trigger",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerResponse"
                }
              }
            },
            "description": "Task triggered"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Trigger a task",
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://runlater.eu",
      "variables": {}
    }
  ],
...
}